CLI-AI-ToolKit/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
google-genai>=0.1.0
ddgs>=1.0.0
